Essai

+++ ExpressVoitures/Data/Models/ReparationModel.cs
@@ -6,7 +6,7 @@
 
         public string Description { get; set; } = null!;
         public decimal Prix { get; set; }
-        public decimal Duree { get; set; }
+        public double Duree { get; set; }
 
         public VoitureModel? Voiture { get; set; }
     }
